chore: don't make AflCmdGen modify files, strictly generate commands only
